+++ src/main/java/ru/yandex/practicum/filmorate/dao/ReviewDbLikeStorage.java
@@ -3,8 +3,8 @@ package ru.yandex.practicum.filmorate.dao;
 import ru.yandex.practicum.filmorate.model.LikeStatus;
 
 public interface ReviewDbLikeStorage {
-    void addLike(long reviewId, Integer userId, LikeStatus status);
+    void addLike(Integer reviewId, Integer userId, LikeStatus status);
 
-    void removeLike(long reviewId, Integer userId, LikeStatus status);
+    void removeLike(Integer reviewId, Integer userId, LikeStatus status);
 
 }
